requirements.txt
aniso8601==8.0.0
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
dj-database-url==0.5.0
Flask==1.1.1
Flask-RESTful==0.3.7
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mpld3==0.3
numpy==1.18.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.10.5
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
Werkzeug==0.16.0
whitenoise==5.0.1
